Use the mode of page hits as a per-host bias.
Here, the bias is the minimum amount of hits the pages that load
a subresource should have for them to be "relevant".

So for a given subresource on a host where the mode of page hits is
5, if sum of hits on pages that loaded that subresource is 5, that
means the subresource probably been loaded by either (a) a collection
of pages with an unusually low number of hits, or (b) a single page,
and in either case we can't really consider that subresource to be
important.

This assumes that the mode really dominates other possible hit values,
which does happen on YouTube, where the mode is 1, and it seems to be
a reasonable assumption for other domains.

- Why not the mean?
We would like the bias to be an integer measure of hits. Also, mean
doesn't work too well for YouTube: because of a few pages with unusually
high hit count, the average gravitates towards 2, which gives some bad
clusters.

- Why not the median?
You got me there, I haven't tried it (yet).