Attempt a new prediction algorithm.
We now no longer restrict ourselves to picking a single cluster containing most
of the resources in a page's last load; instead, we pick the C clusters that
best cover these resources. The set of resources covered by these is the
'resource cover'

Then, from the union of the subclusters in these clusters, we again pick the C
that best cover the resource cover, but this time make sure we never pick too
many resources that are not in the resource cover.

This should greatly increase our 'batting' rate for the subresources
in the last page load, and make the speculativeness of the algorithm (that is,
how many resources we actually pick speculatively vs. the ones that were loaded
on the last page load) configurable.